backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
SQLAlchemy>=2.0
asyncpg
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
aiofiles
greenlet